requirements.txt

numpy==1.23.4
opencv-python-headless==4.8.1.78
scikit-learn==1.1.3
pandas==1.5.1
Pillow==9.2.0
streamlit==1.14.0
